# src/main/resources/openapi3.spec.yaml
openapi: 3.0.3
info:
  title: B-VSA Semestrálny Projekt 2
  description: |
    Špecifikácia REST webových služieb slúžia pre detailný opis požadovanej funkcionality pre Semestrálny Projekt 2 
    v rámci predmetu B-VSA akademický rok 2022/2023.
    
    Pre dopyty na väčšinu služieb (okrem createStudent a createTeacher) je potrebné byť prihlásený. Služby pracujú s Basic
    autentifikáciou. Každý študent a učiteľ je používateľ systému, ktorý sa môže prihlásiť.
    Niektoré služby vyžadujú autorizáciu/oprávnenie pre ich dopyt. Služba, ktorá je obmedzená oprávneniami, to má uvedené
    v popise služby.
    
    Ak služba má vrátiť jeden objekt a nie je nájdený namiesto hodnoty null vráti odpoveď 404. V prípade nečakanej chyby
    má byť odpoveď s kódom 500 a v správe dôvod vzniku chyby (správa zo vzniknutej Exception).
  version: 1.0.0
  license:
    name: Unlicense license
    url: https://unlicense.org
...
servers:
  - url: 'http://localhost:8080'
...
paths:
  /api/students:
    post:
      operationId: createStudent
      tags:
        - student
      summary: Create new student
      description: |
        Vytvorenie nového študenta. Po vytvorení je možné sa prihlásiť ako novo registrovaný študent.
        Služba je verejne dostupná bez prihlásenia.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStudentRequest'
      responses:
        201:
          description: Objekt novo vytvoreného študenta.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
        500:
          $ref: '#/components/responses/ServerError'
    get:
      operationId: getAllStudents
      tags:
        - student
      summary: Get all students
      description: Získanie zoznamu všetkých študentov.
      responses:
        200:
          description: Zoznam všetkých študentov v systéme. Ak v systéme neexistujú žiadny študenti je vrátení prázdny zoznam.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Student'
        401:
          $ref: '#/components/responses/Unauthorized'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/ServerError'
      security:
        - basicAuth: [ ]
  /api/students/{id}:
    get:
      operationId: getStudent
      tags:
        - student
      summary: Get a student by id
      description: Získanie študenta podľa identifikátora.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Objekt študenta so zadaným identifikátorom. V prípade, že taký študent neexistuje tak je vrátená odpoveď 404.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
        401:
          $ref: '#/components/responses/Unauthorized'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/ServerError'
      security:
        - basicAuth: [ ]
    delete:
      operationId: deleteStudent
      tags:
        - student
      summary: Delete a student by id
      description: |
        Vymazanie študenta podľa identifikátora. 
        Vymazať študenta môže iba učiteľ (t.j. prihlásený je učiteľ) alebo prihlásený študent s rovnakým identifikátorom ako v dopyte.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Objekt študenta ktorý bol vymazaný.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/ServerError'
      security:
        - basicAuth: [ ]
  /api/teachers:
    post:
      operationId: createTeacher
      tags:
        - teacher
      summary: Create new teacher
      description: |
        Vytvorenie nového učiteľa. Po vytvorení je možné sa prihlásiť ako novo registrovaný učiteľ.
        Služba je verejne dostupná bez prihlásenia.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTeacherRequest'
      responses:
        201:
          description: Objekt novo vytvoreného učiteľa.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Teacher'
        500:
          $ref: '#/components/responses/ServerError'
    get:
      operationId: getAllTeachers
      tags:
        - teacher
      summary: Get all teachers
      description: Získanie zoznamu všetkých učiteľov.
      responses:
        200:
          description: Zoznam všetkých učiteľov v systéme. Ak v systéme neexistujú žiadny učitelia je vrátený prázdny zoznam.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Teacher'
        401:
          $ref: '#/components/responses/Unauthorized'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/ServerError'
      security:
        - basicAuth: [ ]
  /api/teachers/{id}:
    get:
      operationId: getTeacher
      tags:
        - teacher
      summary: Get a teacher by id
      description: Získanie učiteľa podľa identifikátora.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Objekt učiteľa so zadaným identifikátorom. V prípade, že taký študent neexistuje tak je vrátená odpoveď 404.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Teacher'
        401:
          $ref: '#/components/responses/Unauthorized'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/ServerError'
      security:
        - basicAuth: [ ]
    delete:
      operationId: deleteTeacher
      tags:
        - teacher
      summary: Delete a teacher by id
      description: |
        Vymazanie učiteľa podľa identifikátora.
        Vymazať učiteľa môže iba prihlásený učiteľ, ktorého identifikátor sa zhoduje s poskytnutým identifikátorom v dopyte.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Objekt študenta ktorý bol vymazaný.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Teacher'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/ServerError'
      security:
        - basicAuth: [ ]
...
components:
  schemas:
    Student:
      type: object
      description: Objekt študenta.
      required: [ id, aisId, email ]
      properties:
        id:
          description: Id študenta. V prípade ak je použité AIS ID ako identifikátor študenta tak je táto hodnota rovnaká ako v atribúte aisId.
          type: integer
          format: int64
        aisId:
          description: AIS ID študenta. Zadané pri vytvorení študenta. Musí byť unikátne v rámci všetkých študentov.
          type: integer
          format: int64
        name:
          type: string
        email:
          type: string
          format: email
        year:
          type: integer
        term:
          type: integer
        programme:
          type: string
        thesis:
          $ref: '#/components/schemas/Thesis'
    CreateStudentRequest:
      type: object
      description: Objekt študenta pre vytvorenie nového študenta.
      required: [ aisId, email, name, password ]
      properties:
        aisId:
          description: AIS ID študenta. Musí byť unikátne v rámci všetkých študentov.
          type: integer
          format: int64
        name:
          type: string
        email:
          type: string
          format: email
        password:
          description: Heslo pre prihlásenie študenta. Heslo musí byť zaslané enkódované v BASE64.
          type: string
          format: password
        year:
          type: integer
        term:
          type: integer
        programme:
          type: string
    StudentAltResponse:
      type: object
      description: Objekt študenta s id záverečnej práce namiesto objektu.
      required: [ id, aisId, email ]
      properties:
        id:
          description: Id študenta. V prípade ak je použité AIS ID ako identifikátor študenta tak je táto hodnota rovnaká ako v atribúte aisId.
          type: integer
          format: int64
        aisId:
          description: AIS ID študenta. Zadané pri vytvorení študenta. Musí byť unikátne v rámci všetkých študentov.
          type: integer
          format: int64
        name:
          type: string
        email:
          type: string
          format: email
        year:
          type: integer
        term:
          type: integer
        programme:
          type: string
        thesis:
          description: ID záverečnej práce priradené študentovi. Ak študent nemá priradenú prácu, tento atribút je vynechaný.
          type: integer
          format: int64
    Teacher:
      type: object
      description: Objekt učiteľa.
      required: [ id, aisId, email ]
      properties:
        id:
          description: Id učiteľa. V prípade ak je použité AIS ID ako identifikátor učiteľa tak je táto hodnota rovnaká ako v atribúte aisId.
          type: integer
          format: int64
        aisId:
          description: AIS ID učiteľa. Zadané pri vytvorení učiteľa. Musí byť unikátne v rámci všetkých učiteľov.
          type: integer
          format: int64
        name:
          type: string
        email:
          type: string
          format: email
        institute:
          type: string
        department:
          type: string
        theses:
          type: array
          items:
            $ref: '#/components/schemas/Thesis'
    CreateTeacherRequest:
      type: object
      description: Objekt učiteľa pre vytvorenie nového učiteľa.
      required: [ aisId, email, password, name ]
      properties:
        aisId:
          description: AIS ID učiteľa. Musí byť unikátne v rámci všetkých učiteľov.
          type: integer
          format: int64
        name:
          type: string
        email:
          type: string
          format: email
        password:
          description: Heslo pre prihlásenie učiteľa. Heslo musí byť zaslané enkódované v BASE64.
          type: string
          format: password
        institute:
          type: string
        department:
          type: string
    TeacherAltResponse:
      type: object
      description: Objekt študenta s identifikátormi záverečnej práce namiesto kolekcie objektov.
      required: [ id, aisId, email ]
      properties:
        id:
          description: Id učiteľa. V prípade ak je použité AIS ID ako identifikátor učiteľa tak je táto hodnota rovnaká ako v atribúte aisId.
          type: integer
          format: int64
        aisId:
          description: AIS ID učiteľa. Zadané pri vytvorení učiteľa. Musí byť unikátne v rámci všetkých učiteľov.
          type: integer
          format: int64
        name:
          type: string
        email:
          type: string
          format: email
        institute:
          type: string
        department:
          type: string
        theses:
          type: array
          description: ID záverečných prác vytvorené učiteľom. Ak učiteľ nevytvoril žiadne záverečné práce je tento atribút vynechaný.
          items:
            type: integer
            format: int64
    Thesis:
      type: object
      description: Objekt záverečnej práce.
      required: [ id, registrationNumber, title, department, supervisor, type ]
      properties:
        id:
          type: integer
          format: int64
        registrationNumber:
          type: string
          pattern: ^FEI-\w+$
        title:
          type: string
        description:
          type: string
        department:
          type: string
        supervisor:
          $ref: '#/components/schemas/TeacherAltResponse'
        author:
          $ref: '#/components/schemas/StudentAltResponse'
        publishedOn:
          type: string
          format: date
        deadline:
          type: string
          format: date
        type:
          type: string
          enum:
            - BACHELOR
            - MASTER
            - DISSERTATION
        status:
          type: string
          enum:
            - FREE_TO_TAKE
            - IN_PROGRESS
            - SUBMITTED
...
    Message:
      type: object
      required: [ code, message ]
      description: Objekt správy pre vrátenie informácii o dopyte. Využité najmä v prípade vrátenia odpovedi pri chybe dopytu alebo spracovania.
      properties:
        code:
          type: integer
          pattern: ^\d{3}$
        message:
          type: string
        error:
          type: object
          properties:
            type:
              type: string
            trace:
              type: string
  responses:
    Unauthorized:
      description: |
        Unauthorized request. 
        Dopyt musí byt autorizovaný platnou autentifikačnou schémou. Kód v odpovedi je vždy 401.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
      headers:
        WWW_Authenticate:
          schema:
            type: string
    Forbidden:
      description: |
        Authorized user doesn't have permission for the request.
        Autentifikovaný používateľ nemá oprávnenie na dopyt. Kód v odpovedi je vždy 403.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
    NotFound:
      description: |
        Requested resource was not found.
        Objekt nebolo možné nájsť. Kód v odpovedi je vždy 404.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
    ServerError:
      description: |
        Internal server error.
        Nepredvídateľná chyba na servery pri spracovaní dopytu. Kód v odpovedi je vždy 500.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic